+++ src/components/SortIndicator/CHANGELOG.md
@@ -3,3 +3,5 @@
 - 2025-10-12: Code formatting improvement.
 - 2025-11-07: Support the new hook `useDirection` and remove the old hook. Let
   the `SortIndicator` has its style.
+- 2025-11-15: Add the variant `compact`. The `size` prop is removed -- user
+  should use CSS to set its size.